+++ NEWS.md
@@ -6,6 +6,7 @@
 * Added Estevez_133, Estevez_259, LauxdaCosta_OD_3, and Maertzdorf_15 signatures to the package
 * Added a COVIDsignatures object to the package that can be used to profile COVID-19 gene transcript signatures, thanks to collaborator Dylan Sheerin (WEHI)
 * Added functions to evaluate some signatures using their original models from Johnson lab member Xutao Wang
+* Added pROC option to obtain confidence intervals on AUC values as part of \code{tableAUC()}
 
 # TBSignatureProfiler 0.0.0.9000